third_party/WebKit/LayoutTests/inspector/sources/debugger-async/async-await/async-callstack-async-await1-expected.txt

Tests asynchronous call stacks for async functions.

Set timer for test function.
Captured call stacks in no particular order:
Call stack:
    0) doTestChainedPromises (async-callstack-async-await1.html:100)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await1.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await1.html:21)
    1) timeoutPromise (async-callstack-async-await1.html:9)
    2) doTestChainedPromises (async-callstack-async-await1.html:99)
    3) testFunctionTimeout (async-callstack-async-await1.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await1.html:43)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Call stack:
    0) doTestChainedPromises (async-callstack-async-await1.html:102)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await1.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await1.html:21)
    1) timeoutPromise (async-callstack-async-await1.html:9)
    2) doTestChainedPromises (async-callstack-async-await1.html:101)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await1.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await1.html:21)
    1) timeoutPromise (async-callstack-async-await1.html:9)
    2) doTestChainedPromises (async-callstack-async-await1.html:99)
    3) testFunctionTimeout (async-callstack-async-await1.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await1.html:43)

Call stack:
    0) doTestChainedPromises (async-callstack-async-await1.html:104)
    [Promise.resolve]
    0) doTestChainedPromises (async-callstack-async-await1.html:103)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await1.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await1.html:21)
    1) timeoutPromise (async-callstack-async-await1.html:9)
    2) doTestChainedPromises (async-callstack-async-await1.html:101)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await1.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await1.html:21)
    1) timeoutPromise (async-callstack-async-await1.html:9)
    2) doTestChainedPromises (async-callstack-async-await1.html:99)
    3) testFunctionTimeout (async-callstack-async-await1.html:50)

Call stack:
    0) doTestChainedPromises (async-callstack-async-await1.html:106)
    [Promise.resolve]
    0) doTestChainedPromises (async-callstack-async-await1.html:105)
    [Promise.resolve]
    0) doTestChainedPromises (async-callstack-async-await1.html:103)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await1.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await1.html:21)
    1) timeoutPromise (async-callstack-async-await1.html:9)
    2) doTestChainedPromises (async-callstack-async-await1.html:101)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await1.html:12)

Call stack:
    0) errorCallback (async-callstack-async-await1.html:60)
    1) doTestSettledPromisesRejected (async-callstack-async-await1.html:92)
    [Promise.reject]
    0) doTestSettledPromisesRejected (async-callstack-async-await1.html:89)
    1) testFunctionTimeout (async-callstack-async-await1.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await1.html:43)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Call stack:
    0) promiseCallback (async-callstack-async-await1.html:68)
    1) doTestPromiseConstructor (async-callstack-async-await1.html:66)
    2) testFunctionTimeout (async-callstack-async-await1.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await1.html:43)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Call stack:
    0) thenCallback (async-callstack-async-await1.html:55)
    1) doTestChainedPromises (async-callstack-async-await1.html:107)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await1.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await1.html:21)
    1) timeoutPromise (async-callstack-async-await1.html:9)
    2) doTestChainedPromises (async-callstack-async-await1.html:107)
    [Promise.resolve]
    0) doTestChainedPromises (async-callstack-async-await1.html:105)
    [Promise.resolve]
    0) doTestChainedPromises (async-callstack-async-await1.html:103)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await1.html:12)

Call stack:
    0) thenCallback (async-callstack-async-await1.html:55)
    1) doTestChainedPromisesJSON (async-callstack-async-await1.html:119)
    [Promise.resolve]
    0) doTestChainedPromisesJSON (async-callstack-async-await1.html:118)
    [Promise.resolve]
    0) doTestChainedPromisesJSON (async-callstack-async-await1.html:117)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await1.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await1.html:21)
    1) timeoutPromise (async-callstack-async-await1.html:9)
    2) doTestChainedPromisesJSON (async-callstack-async-await1.html:116)
    3) testFunctionTimeout (async-callstack-async-await1.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await1.html:43)

Call stack:
    0) thenCallback (async-callstack-async-await1.html:55)
    1) doTestPromiseConstructor (async-callstack-async-await1.html:70)
    [Promise.resolve]
    0) doTestPromiseConstructor (async-callstack-async-await1.html:66)
    1) testFunctionTimeout (async-callstack-async-await1.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await1.html:43)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Call stack:
    0) thenCallback (async-callstack-async-await1.html:55)
    1) doTestSettledPromisesResolved (async-callstack-async-await1.html:80)
    [Promise.resolve]
    0) doTestSettledPromisesResolved (async-callstack-async-await1.html:79)
    1) testFunctionTimeout (async-callstack-async-await1.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await1.html:43)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>
